requirements.txt
twikit
python-dotenv
py-cord
colorama